[zetasql-toolkit] Move domain exception into their own packages
