--- tools/zetasql-helper/src/main/java/com/google/pso/zetasql/helper/catalog/bigquery/BigQueryCatalogException.java
@@ -1,9 +1,0 @@
-package com.google.pso.zetasql.helper.catalog.bigquery;
-
-public class BigQueryCatalogException extends RuntimeException {
-
-  public BigQueryCatalogException(String message) {
-    super(message);
-  }
-
-}
