+++ tools/zetasql-helper/src/main/java/com/google/pso/zetasql/helper/catalog/bigquery/BigQueryService.java
@@ -28,6 +28,9 @@ import com.google.cloud.bigquery.Routine;
 import com.google.cloud.bigquery.RoutineId;
 import com.google.cloud.bigquery.Table;
 import com.google.cloud.bigquery.TableId;
+import com.google.pso.zetasql.helper.catalog.bigquery.exceptions.BigQueryCatalogException;
+import com.google.pso.zetasql.helper.catalog.bigquery.exceptions.BigQueryResourceNotFound;
+import com.google.pso.zetasql.helper.catalog.bigquery.exceptions.InvalidBigQueryReference;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
