+++ tools/zetasql-helper/src/main/java/com/google/pso/zetasql/helper/catalog/bigquery/exceptions/BigQueryCatalogException.java
@@ -1,0 +1,11 @@
+package com.google.pso.zetasql.helper.catalog.bigquery.exceptions;
+
+import com.google.pso.zetasql.helper.catalog.exceptions.CatalogException;
+
+public class BigQueryCatalogException extends CatalogException {
+
+  public BigQueryCatalogException(String message) {
+    super(message);
+  }
+
+}
