+++ tools/zetasql-helper/src/main/java/com/google/pso/zetasql/helper/catalog/exceptions/CatalogException.java
@@ -1,0 +1,9 @@
+package com.google.pso.zetasql.helper.catalog.exceptions;
+
+public class CatalogException extends RuntimeException {
+
+  public CatalogException(String message) {
+    super(message);
+  }
+
+}
