+++ tools/zetasql-helper/src/main/java/com/google/pso/zetasql/helper/catalog/spanner/exceptions/SpannerCatalogException.java
@@ -1,0 +1,11 @@
+package com.google.pso.zetasql.helper.catalog.spanner.exceptions;
+
+import com.google.pso.zetasql.helper.catalog.exceptions.CatalogException;
+
+public class SpannerCatalogException extends CatalogException {
+
+  public SpannerCatalogException(String message) {
+    super(message);
+  }
+
+}
